video-processor/requirements.txt
opencv-python>=4.8.0
PyOpenGL>=3.1.7
PyOpenGL-accelerate>=3.1.7
glfw>=2.6.0
numpy>=1.24.0
